Update primes.hs
Rewrote numPrimesUpTo to behave tail recursively so as to improve memory efficiency. go is a helper function for numPrimesUpTo with Int arguments a, which represents the number of primes accumulated so far, and n, which is the number we wish to calculate numPrimesUpTo for.  In go a n, if n = 0, then we return the number of primes we have accumulated, which is a. If n is greater than 0, then if n is a prime we increase the accumulator a by 1 and calculate the number of primes up to n-1. If n is not prime, we do nothing to the accumulator and calculate the number of primes up to n-1. In the implementation of numPrimesUpTo, we partially apply go by setting our accumulator equal to 0 (representing the fact that we have not yet found any primes) and set numPrimesUpTo = go 0. To demonstrate how this works, consider numPrimesUpTo 4 = go 0 4 = go 0 3 = go 1 2 = go 2 1 = go 2 0 = 2, which we know is correct because the only primes less than or equal to 4 are 2 and 3.